# app/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal World Model API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/home": {
      "get": {
        "summary": "Home",
        "description": "What changed since the previous visit, what needs attention, and what the user told us.",
        "operationId": "home_home_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Home"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/visits": {
      "post": {
        "summary": "Record Visit",
        "description": "Called when the app comes to the foreground. Moves the \"since last visit\" marker.",
        "operationId": "record_visit_visits_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Record Visit Visits Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/briefs": {
      "post": {
        "summary": "Generate Brief",
        "operationId": "generate_brief_briefs_post",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "weekly",
              "title": "Period"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/briefs/latest": {
      "get": {
        "summary": "Latest Brief",
        "operationId": "latest_brief_briefs_latest_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/briefs/{brief_id}": {
      "get": {
        "summary": "Brief By Id",
        "description": "A notification's deep link opens this one. Whoever asks must own it.",
        "operationId": "brief_by_id_briefs__brief_id__get",
        "parameters": [
          {
            "name": "brief_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Brief Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/notifications": {
      "get": {
        "summary": "Notifications",
        "operationId": "notifications_notifications_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationView"
                  },
                  "title": "Response Notifications Notifications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/read": {
      "post": {
        "summary": "Mark Notifications Read",
        "operationId": "mark_notifications_read_notifications_read_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Mark Notifications Read Notifications Read Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "post": {
        "summary": "Record Event",
        "operationId": "record_event_events_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Record Event Events Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BriefItem": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ItemKind"
          },
          "assertion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assertion Id"
          },
          "other_assertion_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Assertion Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "predicate": {
            "type": "string",
            "title": "Predicate"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "previous_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Value"
          },
          "due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due"
          },
          "overdue": {
            "type": "boolean",
            "title": "Overdue",
            "default": false
          },
          "effective": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective"
          },
          "is_fact": {
            "type": "boolean",
            "title": "Is Fact"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "evidence_quote": {
            "type": "string",
            "title": "Evidence Quote"
          },
          "other_evidence_quote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Evidence Quote"
          },
          "source_label": {
            "type": "string",
            "title": "Source Label"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "assertion_id",
          "subject",
          "predicate",
          "value",
          "is_fact",
          "confidence",
          "evidence_quote",
          "source_label",
          "observed_at"
        ],
        "title": "BriefItem"
      },
      "BriefView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "covers_since": {
            "type": "string",
            "format": "date-time",
            "title": "Covers Since"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WrittenItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "period",
          "created_at",
          "covers_since",
          "items"
        ],
        "title": "BriefView"
      },
...
      "CommitmentItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "what": {
            "type": "string",
            "title": "What"
          },
          "commitment_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commitment Type"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "committed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed By"
          },
          "committed_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed To"
          },
          "due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due"
          },
          "overdue": {
            "type": "boolean",
            "title": "Overdue"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "review": {
            "type": "string",
            "title": "Review"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "evidence_quote": {
            "type": "string",
            "title": "Evidence Quote"
          },
          "has_conflict": {
            "type": "boolean",
            "title": "Has Conflict"
          },
          "source": {
            "$ref": "#/components/schemas/SourceSummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "what",
          "commitment_type",
          "direction",
          "committed_by",
          "committed_to",
          "due",
          "overdue",
          "status",
          "origin",
          "review",
          "confidence",
          "evidence_quote",
          "has_conflict",
          "source"
        ],
        "title": "CommitmentItem"
      },
...
      "EventIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "EventIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Home": {
        "properties": {
          "since": {
            "type": "string",
            "format": "date-time",
            "title": "Since"
          },
          "what_changed": {
            "items": {
              "$ref": "#/components/schemas/WrittenItem"
            },
            "type": "array",
            "title": "What Changed"
          },
          "what_changed_total": {
            "type": "integer",
            "title": "What Changed Total"
          },
          "needs_attention": {
            "items": {
              "$ref": "#/components/schemas/CommitmentItem"
            },
            "type": "array",
            "title": "Needs Attention"
          },
          "needs_attention_total": {
            "type": "integer",
            "title": "Needs Attention Total"
          },
          "remembered": {
            "items": {
              "$ref": "#/components/schemas/CommitmentItem"
            },
            "type": "array",
            "title": "Remembered"
          },
          "unread_notifications": {
            "type": "integer",
            "title": "Unread Notifications"
          }
        },
        "type": "object",
        "required": [
          "since",
          "what_changed",
          "what_changed_total",
          "needs_attention",
          "needs_attention_total",
          "remembered",
          "unread_notifications"
        ],
        "title": "Home"
      },
...
      "ItemKind": {
        "type": "string",
        "enum": [
          "due_soon",
          "changed",
          "conflict",
          "possible_commitment",
          "consumer",
          "remembered"
        ],
        "title": "ItemKind"
      },
...
      "NotificationView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "deep_link": {
            "type": "string",
            "title": "Deep Link"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "read": {
            "type": "boolean",
            "title": "Read"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "deep_link",
          "created_at",
          "read"
        ],
        "title": "NotificationView"
      },
...
      "SourceSummary": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "sender_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Name"
          },
          "sender_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Address"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "sender_name",
          "sender_address",
          "subject",
          "observed_at"
        ],
        "title": "SourceSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WrittenItem": {
        "properties": {
          "item": {
            "$ref": "#/components/schemas/BriefItem"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "why_it_matters": {
            "type": "string",
            "title": "Why It Matters"
          },
          "suggested_next_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Next Step"
          }
        },
        "type": "object",
        "required": [
          "item",
          "headline",
          "why_it_matters"
        ],
        "title": "WrittenItem"
      }
    }
  }
}